querying by image works

+++ app/src/main/java/com/legec/imgsearch/app/restConnection/dto/Vocabulary.java
@@ -4,13 +4,13 @@ package com.legec.imgsearch.app.restConnection.dto;
 import java.util.List;
 
 public class Vocabulary {
-    private List<List<Double>> vocabulary;
+    private List<List<Float>> vocabulary;
 
-    public List<List<Double>> getVocabulary() {
+    public List<List<Float>> getVocabulary() {
         return vocabulary;
     }
 
-    public void setVocabulary(List<List<Double>> vocabulary) {
+    public void setVocabulary(List<List<Float>> vocabulary) {
         this.vocabulary = vocabulary;
     }
 }
